updating some types

+++ src/depth_buffer.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <assert.h>
 
-depthbuffer* depthbuffer_new(int width, int height)
+depthbuffer* depthbuffer_new(uint16_t width, uint16_t height)
 {
     assert(width > 0 && height > 0);
